Making oauth2 file more user friendly
getpass.getpass() is replaced with input() for the email input and ID input. Secret input is still getpass.getpass() although it can easily be input() because all this info is stored unencrypted in a file on the system. Anyway I'll leave it like that so there is some secrecy to it.