Add support for attaching bytes objects
There are two ways to attach bytes objects:
1. Directly: I.e. `attachments` becomes a list of bytes objects
2. Via a tuple (str, bytes), where the first element is the filename and the second element the file contents.

1. was the first implementation, but I noticed that the filename + suffix was garbled and therefore not user friendly.
Not sure that option should even be supported.

For passing a tuple in 2. it was required to modify the `serialize_object` function to ignore tuples.
Not sure, whether is a backwards incompatible change? Otherwise maybe create a dummy class, like `raw` or `inline`.